fix(tools): extract CargoGridViewer constants to separate TS module
Moves shared types, constants, and utility functions from the Vue SFC
<script> block into constants.ts so they can be properly imported by
TypeScript without vue-shim issues. Fixes tsc and prettier CI failures.
